add rupiah format on planning initial

+++ app/src/main/java/com/example/finpocket/model/PlanItem.kt
@@ -3,7 +3,7 @@ package com.example.finpocket.model
 data class PlanItem(
     val icon: Int,
     val name: String,
-    val percentage: Double,
+    val percentage: Float,
     val nominal: Int
 )
